unify AssociativeBoth and AssociativeEither instances
